+++ redis_array_impl.c
@@ -291,7 +291,7 @@ RedisArray *ra_load_array(const char *name) {
     }
 
     /* create RedisArray object */
-    ra = ra_make_array(hHosts, &z_fun, &z_dist, hPrev, b_index, b_pconnect, (long)l_retry_interval,
+    ra = ra_make_array(hHosts, &z_fun, &z_dist, hPrev, b_index, b_pconnect, l_retry_interval,
                        b_lazy_connect, d_connect_timeout, read_timeout, consistent, algorithm,
                        user, pass);
     if (ra) {
